Factorize feature_detection_generator and semantic_segmentation_generator into a single create_generator function
This commit merge both generator creating functions, both research
problems are now distinguished with a `model` parameter in the new
function. This modification will improve the code in the sense that we
won't have to manage the case distinction in caller
module. Additionnally, all the tests contained in
`./tests/test_generator.py` have been updated.