dataset: Add a new module to manage dataset creation
This commit allows to create easily new image datasets, regarding a
dataset type (`mapillary` or `shapes`), an image size and image
quantities. A particular care is provided to datapaths, so as to
respect some internal path conventions. When the dataset is created
(or if it is already on the file system), the label popularity inside
the training set is printed onto the console.