tests: update feature detection labelling tests
Remove random array generation, and delete a duplicated tests. There
are now `test_feature_detection_labelling_concise`, which tests the
case where labels and array values are the same; and
`test_feature_detection_labelling_sparse` which tests the case where
the array contains values that are not stored in reference label list.