tests: remove useless data sample created for tests
This commit deletes the range of image files (plus .json config files)
that were used during the test. These files are not required any more,
as test fixtures are now used. Temporary data directories and files
are managed during the test verification in /tmp directory, and have
no reason to be in the git repo.