training: fix backup filename when recovering a trained model
As backup filename when sorted by epoch, backup were set as the last
item in the backup repository. However, it seems that the `os.listdir`
method does not provide the files in the alphabetic order;
consequently we do not get the most recent backup. By using `max`
function applied on backup filenames, we ensure that the most recent
backup is chosen (*i.e.* the backup with the largest epoch).