training: setup the seed parameter before generating images fix #45
This commit solves a major issue caused by `None` seed. With the
current implementation, we cannot ensure that each image is linked
with the corresponding labelled version, during data generator
building. To address the issue, we define a seed from the machine time
during execution, and pass it as a parameter to the different
`create_generator` calls.