inference: fix the model initialization by recovering dropout and network architecture from a .json file
In the case where the dropout and the network architecture are not
passed as argument, their default value is None. It means that we must
not use them to initialize a model. However when we want to use the
best available model, the corresponding values must be recovered. They
are in a .json file next to the .h5 model, on the file system. We get
them through a new function in `utils.py` module.