inference: fix the model backup recovering
When recovering a backup in `inference.py`, both handled cases (when a
best model is used, or when training parameters are specified) were
wrong. In the first case, a filename is indicated and we have to
verify if the file exists. In the second case, we are looking into a
instance directory, so we have to check if there is at least one model
backup.