app: modify image path management to introduce ground-truth labels into demo
This commit extends the web app modification to javascript code, to
make the demo page show the ground truth image. At this point, we must
add that symbolic links are fundamental to succeed in displaying the
demo results. There is a `shapes` link that directs to the demo shapes
dataset, either do the `mapillary` symbolic link. Both datasets are
structured with two subdirectories (`images` and `labels`) where
images are picked.