model: update DilatedNet architecture
This commit adapts dilated convolutions (see Yu et al. 2015). As the
output layer has not the same dimension than the input layer, it is
chosen to add some simple transposed convolution layers so as to get
back to the original shape. This feature is experimental, and may be
reverted later if it does not give satisfaction.