app: refactor the image generation and label prediction process
In order to dedicate a specific URL for each image, the web
application is heavily refactored. This commit introduces the
flask_utils_js package so as to generate a new webpage from the
Javascript code, through a given Flask route. From now, the images and
their labels are managed on the server-side, and sent to webpages with
the help of Jinja templates.