training: remove inference step after training
This commit delete testing phase done once the model is trained, as
the inference may be done in a dedicated module, then such an
inference phase is not really useful at this point of the process
(`useful` means having an added value within the project). Moreover
doing inference right after training a model causes memory issues, as
more data needs to be loaded.