[dataset] fix extract_points_from_polygon function
The method does not depend on TanzaniaDataset class, it is moved from
the class definition to a simple function.

Additionnally, the pixel normalization (with respect to min-x and
min-y) has been moved to the function.

And finally there seems to be a coordinates inversion when converting
geometrical coordinates to pixels. This small misunderstanding causes
a major misalignment between images and their labelled version, and it
must be fixed before to train any model.