[datasets] pass the process number as an argument to `populate(.)` functions fix #110
Exploit the number of processes indicated in config.ini or whatever
used config file. By default, consider a single process, and as a
consequence, do not use `Pool(.)` method to keep the execution
simple. This configuration is used during unit tests, as there are
some strange and unresolved bugs when mixing multiprocessing and
pytest (some tests do not pass if there is multiprocessing).