train: merge train.py and paramoptim.py fix #124
`train.py` is removed, and `paramoptim.py` is renamed as
`train.py`. We consider now that training must be done either simple
parameters (in such a case, only one training process is run), or with
list parameters (in this other case, several training processes are
done, and the best model is serialized on the checkpoint subfolder.

As a consequence, this commit fix issue #124 (`train.py` now allows to
store a "best" model).