postprocess: consider shape label during postprocessing fix #131
Before this commit, the polygon rasterization process did not consider
polygon labels (i.e. building types). That reduces the output
significance, as every detected object was of type 1 (with respect to
the dataset glossary).

From now, the polygon types are computed during vectorization process,
and provided as a new argument for the subsequent rasterization
process (that are successive during postprocessing).