datagen: refactor the preprocessed dataset generation for geographic datasets
The CLI arguments '-t', '-v' and '-T' represents amounts of raw
images. While it does not really matter for 'shapes' and 'mapillary'
datasets, it clarifies the way geographic datasets ('aerial' and
'tanzania') are considered, especially in training and validation
scopes.

As a side effect, one needs an additional arguments,
'nb-tiles-per-image', that depicts the amount of tiles that will be
drawn from the original raw images (training and validation
context). Example: if '-t=10' and '--nb-tiles-per-image=10', one will
pick 10 images within the original training set, and try to generate
10 tiles from each of them (same random process as before).